Print just first six characters of Common Form hash
This commit changes the build process to print only the first six
characters of the Common Form fingerprint of the NDA terms in the
agreement's preamble. Corresponding edits in the preamble make clear
that the fingerprint characters printed are just the first characters of
the complete fingerprint.

Six characters can be recognized and verified on sight, and still
represent a statistically strong sample of the complete fingerprint for
content verification.

Reducing the printed fingerprint in this way prevents a nefarious
contractor from so easily:

1. modifying one or more random characters of the fingerprint to weaken
argument that the published terms are validly incorporated by reference

2. making selective edits to the copy of the NDA terms reproduced in
this document signed by the parties

3. arguing that absent effective incorporation by reference, the
modified, printed terms control